Added Pictures, made the card a dynamic activity
You can add more cards by simply adding the word(s) of the object on a
newline in the localCards.txt, and then adding the picture with the SAME
NAME (where spaces are replaced with underscore) and all LOWERCASE.

The current pictures are not optimized so some take a bit of time to
load and are all different sizes.

The log scrolls btw.

Hit start, play (winning does nothing but say so), hit the back btn, hit
start, play again. It's randomized and has no repeats.

CURRENT ISSUE:
If you open every card, then try to open another one it crashes. I want
the usedCards to clear so then it could just randomly cycle again. I
failed at that, but I didn't try very long.

TODO:
1) figure out how to optimize pictures
2) make loading more efficient
3) make a loading screen to prevent starting before loading
4) rake in the millions